Added minimum configuration to be able to deploy to heroku
1. Changed react-scripts start to serve -s build in frontend package.json
and added relevant dependency install commands to help memory usage when
building for heroku dynos with limited RAM.
2. Added a Procfile for Heroku which specifies to allocate a frontend and
backend dyno.
3. Added root level package.json to designate to Heroku that the project
is a nodejs project.
4. Added temporary properties file for Google API key so that frontend can
call Google APIs. This will be removed in the future when we find a way to
pass it in without committing the API key to repo.
5. Added a dev command to frontend/package.json so that we can continue
dev work.