Prevent create reviews page from being edited while logged out
This commit adds logic to render an AI art image and message that prompts
the user to log in if not logged in already. Previously, users were able
to bypass the previous restaurants page and directly access the create
reviews page if they had access to the URL or refreshed the create reviews
page (which logs them out because we don't save cookies) and are able to
edit the text boxes to create a review without being logged in. This leads
to a bug where if they click log in, the page refreshes, their text is
still present, but no updates to the create reviews page have been
recorded. This causes the review to fail to be submitted even though the
tet boxes appear to be populated.