Add User Settings Page
Adds in a user settings page where users can change site-related settings.
For now this page allows you to change your username so that it appears
different to you and others on the site. Additionally, an entry to the
navbar was added so that you can use it to navigate to the page to change
user settings. For now, the GoogleLogin and header on the Critics
components are changed so that the username appears there instead of the
user's name on Google. Further backend work is required such that we no
longer submit the original author's name to the frontend and the backend
no longer returns the original author's name as part of getReviews(),
instead opting to retrieve user names and return them to the front end as
part of its Review object. When a user successfully submits user settings,
a banner will display indicating success or error.

This commit also  refactors the error banner. This is because the component is
very reuseable and only differs by a couple of pieces of information that
can just be passed in as props.

This commit also changes the navbar to use the reactstrap component to
standardize components and prevent weird formatting errors as reactstrap
handles a majority of the formatting.

This commit also updates the google login component to use the username when
avaialble. Also updates the critic page to use the username when
available.