Stop sending author ID when creating a review and turn greeting text black
This commit adds code to stop sending author ID when creating a review
because it's no longer needed with the new usernames changes causing
usernames to be stored separately from reviews. Also adds a visual fix to
the welcome message and turns it black like the other text in the navbar.